NTV-11 refactor location of twitch service files
