+++ src/main/kotlin/com/ngenenius/api/controller/TwitchBetaController.kt
@@ -2,7 +2,7 @@ package com.ngenenius.api.controller
 
 import com.ngenenius.api.model.twitch.StreamDetails
 import com.ngenenius.api.model.twitch.TwitchResponse
-import com.ngenenius.api.service.TwitchStreamsService
+import com.ngenenius.api.service.twitch.TwitchStreamsService
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
